+++ src/OrderStandardEasy/Config.php
@@ -10,10 +10,6 @@
  * @version 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_Ogone_OrderStandardEasy_Config extends Pronamic_WP_Pay_Gateways_Ogone_Config {
-	public function get_payment_server_url() {
-		return 'https://secure.ogone.com/ncol/prod/orderstandard.asp';
-	}
-
 	public function get_gateway_class() {
 		return 'Pronamic_WP_Pay_Gateways_Ogone_OrderStandardEasy_Gateway';
 	}