+++ src/Parameters.php
@@ -7,6 +7,7 @@
  * Company: Pronamic
  *
  * @author Remco Tolsma
+ * @version 1.3.2
  * @since 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_Ogone_Parameters {
@@ -200,4 +201,24 @@ class Pronamic_WP_Pay_Gateways_Ogone_Parameters {
 	 * @var string
 	 */
 	const NC_ERROR_PLUS = 'NCERRORPLUS';
+
+	//////////////////////////////////////////////////
+	// API user parameters
+	//////////////////////////////////////////////////
+
+	/**
+	 * Indicator for the USERID parameter.
+	 *
+	 * @since 1.3.2
+	 * @var string
+	 */
+	const USER_ID = 'USERID';
+
+	/**
+	 * Indicator for the PSWD parameter.
+	 *
+	 * @since 1.3.2
+	 * @var string
+	 */
+	const PASSWORD = 'PSWD';
 }